Adapt RACH procedure for large delay
- Add CL options for additional DURATION_RX_TO_TX and k2 value at UE, additional k2 at gNB
- Apply additional DURATION_RX_TO_TX at UE
- Apply the additional k2 at UE for scheduled msg3 by RAR and scheduled UL transmission by DCL. Re-compute the next Tx frame number
- Apply the additional k2 at gNB for scheduled msg3 and scheduled UL transmission. Re-compute the next Tx frame number
- Increase the RA window counter for reception of RAR
- Fix the data type for larger slot number in function get_samples_slot_timestamp
- Example for a one-way delay of 120ms: run gNB with "-P 7372800 --gnb_k2 480", run UE with "-P 7372800 -A 14745600 --ue_slot_Rx_Tx 480 --ue_k2 480"

+++ openair1/PHY/INIT/nr_parms.c
@@ -203,10 +203,10 @@ uint32_t get_samples_slot_timestamp(int slot, const NR_DL_FRAME_PARMS *fp, uint8
   uint32_t samp_count = 0;
 
   if(!sl_ahead) {
-    for(uint8_t idx_slot = 0; idx_slot < slot; idx_slot++)
+    for(uint16_t idx_slot = 0; idx_slot < slot; idx_slot++)
       samp_count += fp->get_samples_per_slot(idx_slot, fp);
   } else {
-    for(uint8_t idx_slot = slot; idx_slot < slot+sl_ahead; idx_slot++)
+    for(uint16_t idx_slot = slot; idx_slot < slot+sl_ahead; idx_slot++)
       samp_count += fp->get_samples_per_slot(idx_slot, fp);
   }
   return samp_count;
